refactor: Add some comments and rename to integtest
