+++ consensus/integtest/listener.go
@@ -1,4 +1,4 @@
-package integ
+package integtest
 
 import (
 	"github.com/NethermindEth/juno/consensus/types"
